# Faster parallel download by using custom IO thread pool