Before
```
real    2m 55.19s
user    0m 10.79s
sys     0m 3.04s
```

After
```
real	1m 7.26s
user	0m 10.53s
sys     0m 3.61s
```

Since this is mostly IO-bound we can go crazy with a custom thread pool.