@casz 2019-08-07
When I initially tried these changes in #49 and now with this PR 
I was on Windows laptop 4 cores (8 threads) running inside docker where we are CPU bound to 2 core.
Still takes around 3 minutes.

However, with my desktop with 8 cores (16 threads) running Linux this PR shaves 15-20 seconds off. From 1m 22s to 1m 5s.

Running on my Windows laptop outside of Docker the difference is 30 seconds.
From 1m 55s to 1m 22s!